sparktest/spark-requirements.txt
flytekit==1.2.4
pyspark==3.2.1
flytekitplugins-spark==1.2.4